#38. Refactor (GamePlayersBrowser.vue): Move the file
