+++ src/entities/game-players/GamePlayersBrowser.vue
